Create Q1.cpp
. Implement the Shortest Job First CPU scheduling algorithm. Read
inputs from a file (format given below).
Inputs:
The program will read from an input file containing a list of
processes along with other data require for
scheduling.
The input file will look like this:
P1 0 20.0
P2 2 15.0
P3 6 27.0
p4 4 36.0
• The file containing the information on the processes will have
each process on a separate line. The
processes will be in the file in the order in which they arrive at the
OS.
• Each line will have a process name that will be a string.
• Following the name will be the arrival time of the process
• Following arrival time will be the total burst time.
Outputs:
The program must print out the time taken by each process to
complete (turnaround time) and the wait
time and compute the average turnaround time for all processes.