Create Q1.java
Given 2 process A and B executing simultaneously, both the process
A and B
share a common, fixed-size array. Process A place an element in the
array. If
array is already full then the process A will have to wait for an
empty slot in
the array. On the other hand, process B removes an item from the
array. If array
is already empty, then the process B will have to wait for an item in
array.
Implement Peterson’s Algorithm for the two processes using shared
memory
such that there is mutual exclusion between them.